* Fix: Wordfence no longer can appear on sub-sites on multi-site installs, only on the network admin panel. * Fix: Wordfence dashboard widget only can appear on network admin dashboard in multi-site installs. * Fix: No more multiple scheduled scans on multi-site. * Fix: Fixed mixed-protocol warning if you're using SSL and Wordfence - our static assets are loaded without specifying protocol now. * Fix: Fixed issue where non-existent users were shown in dashboard widget and email summary as valid users. * Fix: Removed /e modifier in preg_replace for Diff_Renderer_Html_Array::formatLines since it is deprecated in PHP 5.5. * Fix: Removed ssl_verify => false from wp_remote_post connectivity test since some versions of cURL will throw an error since WordPress uses their own certificate bundle. * Fix: Fixed bug with activity report email date range (was one week ahead). * Fix: Removed email summary report from cron on deactivation. * Fix: Fixed an off-by-one bug in wfDirectoryIterator for maximum total files and max files per directory. * Fix: Updated our browser data to fix an issue that caused newer browsers to appear in live traffic with version 0.0. * Improvement: Updated the country database used for country blocking to April 2015 version. * Improvement: Added an additional check for disabling script execution in the uploads directory that the .htaccess file actually contains our protection code before removing it. * Improvement: Paused Live Traffic ajax request when the window/document loses focus to reduce server load. * Improvement: Better error handling when making API calls to noc1 to help our support personell help you. * Improvement: Added locked out IP's and IP's restricted through advanced blocking to the blocked IP log for dashboard and email summary. * Improvement: Excluded whitelisted IP's from dashboard and widget email summary.
git-svn-id: https://plugins.svn.wordpress.org/wordfence/tags/5.3.12@1142353 b8457f37-d9ea-0310-8a92-e5e31aec5664

+++ lib/wfConfig.php
@@ -789,8 +789,11 @@ class wfConfig {
 		return true;
 	}
 	public static function enableAutoUpdate(){
-		wfConfig::set('autoUpdate', '1');	
-		wp_schedule_event(time(), 'daily', 'wordfence_daily_autoUpdate');
+		wfConfig::set('autoUpdate', '1');
+		wp_clear_scheduled_hook('wordfence_daily_autoUpdate');
+		if (is_main_site()) {
+			wp_schedule_event(time(), 'daily', 'wordfence_daily_autoUpdate');
+		}
 	}
 	public static function disableAutoUpdate(){
 		wfConfig::set('autoUpdate', '0');	
@@ -857,11 +860,12 @@ Options -ExecCGI
 		$upload_dir = wp_upload_dir();
 		return $upload_dir['basedir'] . '/.htaccess';
 	}
-	
+
 	/**
 	 * Add/Merge .htaccess file in the uploads directory to prevent code execution.
 	 *
 	 * @return bool
+	 * @throws wfConfigException
 	 */
 	public static function disableCodeExecutionForUploads() {
 		$uploads_htaccess_file_path = self::_uploadsHtaccessFilePath();
@@ -880,27 +884,32 @@ Options -ExecCGI
 		}
 		return true;
 	}
-	
+
 	/**
 	 * Remove script execution protections for our the .htaccess file in the uploads directory.
 	 *
 	 * @return bool
+	 * @throws wfConfigException
 	 */
 	public static function removeCodeExecutionProtectionForUploads() {
 		$uploads_htaccess_file_path = self::_uploadsHtaccessFilePath();
 		if (file_exists($uploads_htaccess_file_path)) {
 			$htaccess_contents = file_get_contents($uploads_htaccess_file_path);
-			$htaccess_contents = str_replace(self::$_disable_scripts_htaccess, '', $htaccess_contents);
-			
-			$error_message = "Unable to remove code execution protections applied to the .htaccess file in the uploads directory.  Please check your permissions on that file.";
-			if (strlen(trim($htaccess_contents)) === 0) {
-				// empty file, remove it
-				if (!@unlink($uploads_htaccess_file_path)) {
+
+			// Check that it is in the file
+			if (strpos($htaccess_contents, self::$_disable_scripts_htaccess) !== false) {
+				$htaccess_contents = str_replace(self::$_disable_scripts_htaccess, '', $htaccess_contents);
+
+				$error_message = "Unable to remove code execution protections applied to the .htaccess file in the uploads directory.  Please check your permissions on that file.";
+				if (strlen(trim($htaccess_contents)) === 0) {
+					// empty file, remove it
+					if (!@unlink($uploads_htaccess_file_path)) {
+						throw new wfConfigException($error_message);
+					}
+
+				} elseif (@file_put_contents($uploads_htaccess_file_path, $htaccess_contents, LOCK_EX) === false) {
 					throw new wfConfigException($error_message);
 				}
-				
-			} elseif (@file_put_contents($uploads_htaccess_file_path, $htaccess_contents, LOCK_EX) === false) {
-				throw new wfConfigException($error_message);
 			}
 		}
 		return true;
